# services/api/openapi.yaml
openapi: 3.0.3
info:
  title: tyr.bet API
  version: 1.0.0
  description: "Testnet-only. Session auth = signed `tyr_session` cookie from the passkey routes. Agent routes are signed with the agent key (x-tyr-agent / x-tyr-timestamp / x-tyr-signature, EIP-191 over method, path, timestamp, sha256(body)); paid agent routes answer 402 with an MPP challenge. Live events: GET /ws (see WsEvent)."
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: tyr_session
    agent:
      type: apiKey
      in: header
      name: x-tyr-signature
    admin:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          description: machine-readable reason (e.g. limit, balance, region)
        issues:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
            required:
              - path
              - message
            additionalProperties: false
          description: request validation failures
      required:
        - error
      additionalProperties: false
      description: Error
...
    Market:
      type: object
      properties:
        outcome:
          type: integer
        name:
          type: string
        description:
          type: string
        template:
          type: object
          additionalProperties:
            type: string
          nullable: true
        sides:
          type: array
          items:
            type: string
        resolvesAt:
          type: string
          format: date-time
        "yes":
          type: object
          properties:
            coin:
              type: string
            bid:
              type: number
            ask:
              type: number
            mid:
              type: number
          required:
            - coin
            - bid
            - ask
            - mid
          additionalProperties: false
      required:
        - outcome
        - name
        - description
        - template
        - sides
        - resolvesAt
        - "yes"
      additionalProperties: false
...
paths:
...
  /api/bets:
    post:
      tags:
        - bets
      description: Place a bet (Flow A). 402 = loss limit exceeded on-chain, 409 = insufficient confidential balance. Replaying an idempotencyKey returns the same order.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                outcome:
                  type: integer
                  minimum: 0
                side:
                  type: string
                  enum:
                    - "yes"
                    - "no"
                stakeUsd:
                  type: number
                  exclusiveMinimum: true
                  minimum: 0
                  maximum: 1000
                idempotencyKey:
                  type: string
                  minLength: 8
                  maxLength: 128
                maxPrice:
                  type: number
                  exclusiveMinimum: true
                  minimum: 0
                  exclusiveMaximum: true
                  maximum: 1
              required:
                - outcome
                - side
                - stakeUsd
                - idempotencyKey
              additionalProperties: false
      security:
        - session: []
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "402":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags:
        - bets
      security:
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  bets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
                required:
                  - bets
                additionalProperties: false
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/bets/{id}:
    get:
      tags:
        - bets
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: id
          required: true
      security:
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/agent/sessions:
    post:
      tags:
        - agent sessions
      description: Register an agent key with a spend cap; the passkey then authorizes it on-chain.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                agentAddress:
                  type: string
                  pattern: ^0x[0-9a-fA-F]{40}$
                capUsd:
                  type: number
                  exclusiveMinimum: true
                  minimum: 0
                  maximum: 10000
                period:
                  type: string
                  enum:
                    - day
                    - week
                  default: day
                name:
                  type: string
                  maxLength: 64
                expiresInDays:
                  type: integer
                  minimum: 1
                  maximum: 90
                  default: 30
              required:
                - agentAddress
                - capUsd
              additionalProperties: false
      security:
        - session: []
      responses:
        "201":
          description: Arguments for viem/tempo Actions.accessKey.authorize, signed by the passkey
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessKey:
                    type: object
                    properties:
                      accessKeyAddress:
                        type: string
                        pattern: ^0x[0-9a-fA-F]{40}$
                      keyType:
                        type: string
                        enum:
                          - p256
                          - secp256k1
                    required:
                      - accessKeyAddress
                      - keyType
                    additionalProperties: false
                  expiry:
                    type: integer
                    description: unix seconds
                  limits:
                    type: array
                    items:
                      type: object
                      properties:
                        token:
                          type: string
                          pattern: ^0x[0-9a-fA-F]{40}$
                        limit:
                          type: string
                          description: token base units (6 decimals)
                        period:
                          type: integer
                          description: seconds
                      required:
                        - token
                        - limit
                        - period
                      additionalProperties: false
                  sessionId:
                    type: string
                required:
                  - accessKey
                  - expiry
                  - limits
                  - sessionId
                additionalProperties: false
                description: Arguments for viem/tempo Actions.accessKey.authorize, signed by the passkey
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags:
        - agent sessions
      security:
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                          nullable: true
                        agentAddress:
                          type: string
                        capUsd:
                          type: number
                        spentUsd:
                          type: number
                        period:
                          type: string
                        expiresAt:
                          type: string
                          format: date-time
                        revokedAt:
                          type: string
                          format: date-time
                          nullable: true
                        authorizedTx:
                          type: string
                          nullable: true
                        active:
                          type: boolean
                        remainingUsd:
                          type: number
                          nullable: true
                      required:
                        - id
                        - name
                        - agentAddress
                        - capUsd
                        - spentUsd
                        - period
                        - expiresAt
                        - revokedAt
                        - authorizedTx
                        - active
                        - remainingUsd
                      additionalProperties: false
                required:
                  - sessions
                additionalProperties: false
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/agent/sessions/{id}/confirm:
    put:
      tags:
        - agent sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                txHash:
                  type: string
                  pattern: ^0x[0-9a-fA-F]{64}$
              required:
                - txHash
              additionalProperties: false
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: id
          required: true
      security:
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                  remainingUsd:
                    type: number
                required:
                  - sessionId
                  - remainingUsd
                additionalProperties: false
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/agent/sessions/{id}:
    delete:
      tags:
        - agent sessions
      description: Stops tyr serving the agent; revoke the on-chain key with the passkey.
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: id
          required: true
      security:
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                  revoked:
                    type: boolean
                    enum:
                      - true
                required:
                  - sessionId
                  - revoked
                additionalProperties: false
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/agent/me:
    get:
      tags:
        - agent
      security:
        - agent: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                    nullable: true
                  agentAddress:
                    type: string
                  capUsd:
                    type: number
                  spentUsd:
                    type: number
                  period:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
                  revokedAt:
                    type: string
                    format: date-time
                    nullable: true
                  authorizedTx:
                    type: string
                    nullable: true
                  active:
                    type: boolean
                  owner:
                    type: string
                    pattern: ^0x[0-9a-fA-F]{40}$
                  remainingUsd:
                    type: number
                required:
                  - id
                  - name
                  - agentAddress
                  - capUsd
                  - spentUsd
                  - period
                  - expiresAt
                  - revokedAt
                  - authorizedTx
                  - active
                  - owner
                  - remainingUsd
                additionalProperties: false
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/agent/markets:
    get:
      tags:
        - agent
      security:
        - agent: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  markets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Market"
                  priceUsd:
                    type: object
                    properties:
                      data:
                        type: number
                    required:
                      - data
                    additionalProperties: false
                required:
                  - markets
                  - priceUsd
                additionalProperties: false
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...